Add editor safety command to clippy
As people often ask if it's safe to send editor links, it would be nice to have a command we can use to explain it to them. Though it is rather wordy, this is due to the fact that people who are concerned about it are usually more prepared to read, and since to them they have valid concerns I felt it important to clearly state why this is not a problem.